sysdig: refactor into two parsers
