+++ docs/content/en/connecting_your_tools/parsers/file/sysdig_reports.md
@@ -2,14 +2,11 @@
 title: "Sysdig Vulnerability Reports"
 toc_hide: true
 ---
-Legacy Scanning Engine:
-
 Import CSV report files from Sysdig or a Sysdig UI JSON Report
 Parser will accept Pipeline, Registry and Runtime reports created from the UI
-More information available at [our reporting docs page](https://docs.sysdig.com/en/docs/sysdig-secure/vulnerabilities/reporting)
+More information available at [sysdig reporting docs page](https://docs.sysdig.com/en/docs/sysdig-secure/vulnerabilities/reporting)
 
-Vulnerability Management Engine:
-Import CSV and JSON files creared by Sysdig CLI scanner.
+Import CSV and JSON files created by Sysdig CLI scanner.
 
 
 ### Sample Scan Data
