+++ unittests/scans/sysdig_reports/cli/sysdig_reports_many_vul.csv
@@ -1,0 +1,32 @@
+CVE ID,CVE Severity,CVSS Score,CVSS Score Version,Package Name,Package Version,Package Path,Package Type,Fix Version,CVE URL,Vuln Disclosure Date,Vuln Fix Date,Risk Accepted,EPSS Score
+CVE-2013-7459,Critical,9.8,3.0,pycrypto,2.6.1,/usr/lib/python2.7/dist-packages/pycrypto-2.6.1.egg-info,python,,https://nvd.nist.gov/vuln/detail/CVE-2013-7459,2013-12-22T00:00:00Z,,false,0.00587
+CVE-2019-6446,Critical,9.8,3.0,numpy,1.12.1,/usr/lib/python2.7/dist-packages/numpy-1.12.1.egg-info/PKG-INFO,python,v1.16.1,https://nvd.nist.gov/vuln/detail/CVE-2019-6446,2019-01-16T00:00:00Z,,false,0.03444
+CVE-2022-29361,Critical,9.8,3.1,Werkzeug,1.0.1,/usr/local/lib/python2.7/dist-packages/Werkzeug-1.0.1.dist-info/METADATA,python,v2.1.1,https://nvd.nist.gov/vuln/detail/CVE-2022-29361,2022-05-24T00:00:00Z,,false,0.00101
+CVE-2024-6345,Critical,9.8,3.0,setuptools,33.1.1,/usr/lib/python2.7/dist-packages/setuptools-33.1.1.egg-info/PKG-INFO,python,v70.0.0,,2024-07-14T00:00:00Z,2024-05-21T00:00:00Z,false,0.00043
+CVE-2017-12852,High,7.5,3.0,numpy,1.12.1,/usr/lib/python2.7/dist-packages/numpy-1.12.1.egg-info/PKG-INFO,python,v1.16.1,https://nvd.nist.gov/vuln/detail/CVE-2017-12852,2017-08-15T00:00:00Z,,false,0.0017
+CVE-2018-6594,High,7.5,3.0,pycrypto,2.6.1,/usr/lib/python2.7/dist-packages/pycrypto-2.6.1.egg-info,python,,https://nvd.nist.gov/vuln/detail/CVE-2018-6594,2018-02-01T00:00:00Z,,false,0.00292
+CVE-2019-12761,High,7.5,3.0,pyxdg,0.25,/usr/share/pyshared/pyxdg-0.25.egg-info,python,v0.26.0,https://nvd.nist.gov/vuln/detail/CVE-2019-12761,2013-02-05T00:00:00Z,2018-02-02T00:00:00Z,false,0.01209
+CVE-2019-20916,High,7.5,3.1,pip,9.0.1,/usr/lib/python2.7/dist-packages/pip-9.0.1.egg-info/PKG-INFO,python,v19.2.0,https://nvd.nist.gov/vuln/detail/CVE-2019-20916,2019-04-16T00:00:00Z,2019-07-22T00:00:00Z,false,0.00615
+CVE-2022-40898,High,7.5,3.1,wheel,0.29.0,/usr/lib/python2.7/dist-packages/wheel-0.29.0.egg-info/PKG-INFO,python,v0.38.1,https://nvd.nist.gov/vuln/detail/CVE-2022-40898,2022-10-20T00:00:00Z,2022-10-20T00:00:00Z,false,0.00374
+CVE-2023-0286,High,7.4,3.1,cryptography,1.7.1,/usr/lib/python2.7/dist-packages/cryptography-1.7.1.egg-info/PKG-INFO,python,v39.0.1,https://nvd.nist.gov/vuln/detail/CVE-2023-0286,2023-02-07T00:00:00Z,2023-02-07T00:00:00Z,false,0.00422
+CVE-2023-25577,High,7.5,3.1,Werkzeug,1.0.1,/usr/local/lib/python2.7/dist-packages/Werkzeug-1.0.1.dist-info/METADATA,python,v2.1.1,https://nvd.nist.gov/vuln/detail/CVE-2023-25577,2023-02-14T00:00:00Z,2023-02-14T00:00:00Z,false,0.0023
+CVE-2023-30861,High,7.5,3.1,Flask,1.1.2,/usr/local/lib/python2.7/dist-packages/Flask-1.1.2.dist-info/METADATA,python,v2.2.5,https://nvd.nist.gov/vuln/detail/CVE-2023-30861,2023-05-01T00:00:00Z,2023-05-01T00:00:00Z,false,0.00228
+CVE-2023-46136,High,7.5,3.1,Werkzeug,1.0.1,/usr/local/lib/python2.7/dist-packages/Werkzeug-1.0.1.dist-info/METADATA,python,v2.1.1,https://nvd.nist.gov/vuln/detail/CVE-2023-46136,2023-10-24T00:00:00Z,2023-10-24T00:00:00Z,false,0.00095
+CVE-2023-50782,High,7.5,3.1,cryptography,1.7.1,/usr/lib/python2.7/dist-packages/cryptography-1.7.1.egg-info/PKG-INFO,python,v39.0.1,https://nvd.nist.gov/vuln/detail/CVE-2023-50782,2020-10-24T00:00:00Z,,false,0.00154
+CVE-2024-34069,High,7.5,3.0,Werkzeug,1.0.1,/usr/local/lib/python2.7/dist-packages/Werkzeug-1.0.1.dist-info/METADATA,python,v2.1.1,,2024-05-06T00:00:00Z,2024-05-06T00:00:00Z,false,0.00045
+CVE-2024-3651,High,7.5,3.1,idna,2.2,/usr/lib/python2.7/dist-packages/idna-2.2.egg-info/PKG-INFO,python,v3.7.0,https://nvd.nist.gov/vuln/detail/CVE-2024-3651,2024-04-11T00:00:00Z,2024-04-11T00:00:00Z,false,0.00046
+CVE-2024-49767,High,7.5,3.1,Werkzeug,1.0.1,/usr/local/lib/python2.7/dist-packages/Werkzeug-1.0.1.dist-info/METADATA,python,v2.1.1,https://nvd.nist.gov/vuln/detail/CVE-2024-49767,2024-10-25T00:00:00Z,2024-10-25T00:00:00Z,false,0.00056
+CVE-2024-56326,High,8.1,3.0,Jinja2,2.11.2,/usr/local/lib/python2.7/dist-packages/Jinja2-2.11.2.dist-info/METADATA,python,v3.1.5,,2024-12-19T00:00:00Z,2024-12-21T00:00:00Z,false,0.00129
+CVE-2020-25659,Medium,5.9,3.1,cryptography,1.7.1,/usr/lib/python2.7/dist-packages/cryptography-1.7.1.egg-info/PKG-INFO,python,v39.0.1,https://nvd.nist.gov/vuln/detail/CVE-2020-25659,2020-10-24T00:00:00Z,,false,0.00149
+CVE-2020-28493,Medium,5.3,3.1,Jinja2,2.11.2,/usr/local/lib/python2.7/dist-packages/Jinja2-2.11.2.dist-info/METADATA,python,v3.1.5,https://nvd.nist.gov/vuln/detail/CVE-2020-28493,2021-01-31T00:00:00Z,2021-01-31T00:00:00Z,false,0.00235
+CVE-2021-3572,Medium,5.7,3.1,pip,9.0.1,/usr/lib/python2.7/dist-packages/pip-9.0.1.egg-info/PKG-INFO,python,v19.2.0,https://nvd.nist.gov/vuln/detail/CVE-2021-3572,2021-04-24T00:00:00Z,2021-04-24T00:00:00Z,false,0.00091
+CVE-2021-41495,Medium,5.3,3.1,numpy,1.12.1,/usr/lib/python2.7/dist-packages/numpy-1.12.1.egg-info/PKG-INFO,python,v1.16.1,https://nvd.nist.gov/vuln/detail/CVE-2021-41495,2021-05-19T00:00:00Z,2022-02-03T00:00:00Z,false,0.00071
+CVE-2021-41496,Medium,5.5,3.1,numpy,1.12.1,/usr/lib/python2.7/dist-packages/numpy-1.12.1.egg-info/PKG-INFO,python,v1.16.1,https://nvd.nist.gov/vuln/detail/CVE-2021-41496,2021-05-13T00:00:00Z,2020-06-20T00:00:00Z,false,0.00042
+CVE-2022-40897,Medium,5.9,3.1,setuptools,33.1.1,/usr/lib/python2.7/dist-packages/setuptools-33.1.1.egg-info/PKG-INFO,python,v70.0.0,https://nvd.nist.gov/vuln/detail/CVE-2022-40897,2022-11-02T00:00:00Z,2022-11-04T00:00:00Z,false,0.00652
+CVE-2024-0727,Medium,5.5,3.1,cryptography,1.7.1,/usr/lib/python2.7/dist-packages/cryptography-1.7.1.egg-info/PKG-INFO,python,v39.0.1,https://nvd.nist.gov/vuln/detail/CVE-2024-0727,2024-01-22T00:00:00Z,2024-01-30T00:00:00Z,false,0.00228
+CVE-2024-22195,Medium,6.1,3.1,Jinja2,2.11.2,/usr/local/lib/python2.7/dist-packages/Jinja2-2.11.2.dist-info/METADATA,python,v3.1.5,https://nvd.nist.gov/vuln/detail/CVE-2024-22195,2024-01-11T00:00:00Z,2024-05-06T00:00:00Z,false,0.00173
+CVE-2024-34064,Medium,6.1,3.0,Jinja2,2.11.2,/usr/local/lib/python2.7/dist-packages/Jinja2-2.11.2.dist-info/METADATA,python,v3.1.5,,2024-01-11T00:00:00Z,2024-05-06T00:00:00Z,false,0.00044
+CVE-2014-1624,Low,3.3,2.0,pyxdg,0.25,/usr/share/pyshared/pyxdg-0.25.egg-info,python,v0.26.0,https://nvd.nist.gov/vuln/detail/CVE-2014-1624,2013-01-21T06:00:00Z,,false,0.00042
+CVE-2023-23934,Low,3.5,3.1,Werkzeug,1.0.1,/usr/local/lib/python2.7/dist-packages/Werkzeug-1.0.1.dist-info/METADATA,python,v2.1.1,https://nvd.nist.gov/vuln/detail/CVE-2023-23934,2023-02-14T00:00:00Z,2023-02-14T00:00:00Z,false,0.00064
+CVE-2023-5752,Low,3.3,3.1,pip,9.0.1,/usr/lib/python2.7/dist-packages/pip-9.0.1.egg-info/PKG-INFO,python,v19.2.0,https://nvd.nist.gov/vuln/detail/CVE-2023-5752,2023-10-01T00:00:00Z,2023-10-15T00:00:00Z,false,0.00048
+CVE-2024-49766,Low,3.7,3.0,Werkzeug,1.0.1,/usr/local/lib/python2.7/dist-packages/Werkzeug-1.0.1.dist-info/METADATA,python,v2.1.1,,2024-10-25T00:00:00Z,2024-10-25T00:00:00Z,false,0.00045
